[floating-ui] Deduplicate DisabledIndices type and grid row test helper
